+++ src/builder.rs
@@ -116,6 +116,8 @@ impl<T: PartialEq> FormBuilder<T> {
             spacing: 2,
             preview: 2,
             scrollbar: false,
+            pinnable: false,
+            searchable: false,
         }
     }